+++ config/media.php
@@ -1,0 +1,7 @@
+<?php
+
+return [
+
+    'use_x_sendfile' => env('MEDIA_USE_X_SENDFILE', false),
+
+];